Added set function to QuizContext and consolidated files
