--- src/pages/Quiz/default-quiz-context.js
@@ -1,11 +1,0 @@
-const defaultQuizContext = {
-    data: {
-        category: "HTML/CSS",
-        difficulty: "Beginner",
-        type: "Any",
-        time: "1 month",
-        hours: "2 hours",
-    },
-};
-
-export default defaultQuizContext;
