--- src/pages/Quiz/quiz-context.js
@@ -1,6 +1,0 @@
-import { createContext } from "react";
-import defaultQuizContext from "./default-quiz-context";
-
-const QuizContext = createContext(defaultQuizContext);
-
-export default QuizContext;
